refactor: split astro_features into own module
